Delete cheksum from class variables and minor convetion fixes
Changes the function Checksum to return a integer, leaving the calculation of the
number at the time of encoding, therefore deleting checksum as a variable. Changes
other functions accordingly. Update Gtest cases to mirror changes in the design of the
data variable, now it is not assumed to be time, but a general vector of bytes, that
happens to store time in this specific implementation of a TCP server for ping.